backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
strawberry-graphql
httpx
requests
sse-starlette
